refactor(services): reorganize usecase/service into lifecycle/ and query/ sub-packages
- domain/service.go → domain/service_query.go to mirror service_lifecycle.go
- usecase/service/{lifecycle,query,status}.go → usecase/service/lifecycle/ and usecase/service/query/
- Rename Query struct → Reader in query package
- Split mock_test.go per package (each keeps only its relevant mocks)
