+++ services/usecase/service/lifecycle/lifecycle.go
@@ -1,4 +1,4 @@
-package service
+package lifecycle
 
 import (
 	"context"
