+++ services/usecase/service/lifecycle/lifecycle_test.go
@@ -1,4 +1,4 @@
-package service
+package lifecycle
 
 import (
 	"context"
