+++ services/usecase/service/query/status.go
@@ -1,4 +1,4 @@
-package service
+package query
 
 import (
 	"github.com/onyxia-datalab/onyxia-backend/services/domain"
